requirements.txt
confluent_kafka==2.5.0
pyspark==3.2.1
python-dotenv==1.0.1
avro-python3==1.10.2
opencv-python==4.10.0
minio==7.2.7
pillow==10.4.0
